+++ src/components/LayerCollection/LayerPO.js
@@ -10,7 +10,10 @@ import {
 
 import { getUserNameById } from '../../common/utils';
 import { updateOrderResource } from '../Utils/orderResource';
-import { showUpdateOrderError } from '../Utils/order';
+import {
+  showUpdateOrderError,
+  parseOngoingOrderInfo,
+} from '../Utils/order';
 import POForm from '../PurchaseOrder/POForm';
 import { UpdateOrderErrorModal } from '../PurchaseOrder/UpdateOrderErrorModal';
 
@@ -63,8 +66,10 @@ class LayerPO extends Component {
 
   updatePO = (order) => {
     const { parentMutator, onCancel } = this.props;
+    const getOngoingOrderInfo = parseOngoingOrderInfo(order);
 
-    updateOrderResource(order, parentMutator.records)
+    console.log('getOngoingOrderInfo', getOngoingOrderInfo);
+    updateOrderResource(getOngoingOrderInfo, parentMutator.records)
       .then(() => onCancel())
       .catch(async e => {
         await showUpdateOrderError(e, this.callout, this.openOrderErrorModalShow);