+++ src/components/PurchaseOrder/ongoingOgderInfo/index.js
@@ -1,0 +1,4 @@
+import OngoingOrderInfoView from './OngoingOrderInfoView';
+import OngoingInfoForm from './OngoingInfoForm';
+
+export { OngoingOrderInfoView, OngoingInfoForm };