+++ src/components/Utils/order/index.js
@@ -1,3 +1,4 @@
 export { default as showUpdateOrderError, ERROR_CODES } from './showUpdateOrderError';
+export { parseOngoingOrderInfo } from './parseOngoingOrderInfo';
 
 export default {};