+++ test/bigtest/tests/edit-order-test.js
@@ -24,7 +24,7 @@ describe('OrderEditPage', function () {
 
   it('displays Edit Order form', function () {
     expect(orderEditPage.$root).to.exist;
-    expect(orderEditPage.renewalsAccordion).to.be.false;
+    expect(orderEditPage.isOngoingInfoOpen).to.be.false;
     expect(orderEditPage.addNoteButton.isButton).to.be.true;
     expect(orderEditPage.notes().length).to.be.equal(1);
   });
@@ -33,13 +33,15 @@ describe('OrderEditPage', function () {
     expect(orderEditPage.hasTemplateField).to.be.true;
   });
 
-  describe('Select order type Ongoing', function () {
+  describe.only('Select order type Ongoing', function () {
     beforeEach(async function () {
       await orderEditPage.orderTypeSelect.select(ORDER_TYPE.ongoing);
+      await orderEditPage.renewalsAccordion.isSubscription();
     });
 
-    it('displays Renewals Accordion', function () {
-      expect(orderEditPage.renewalsAccordion).to.be.true;
+    it('displays Ongoing Order Info Accordion', function () {
+      expect(orderEditPage.isOngoingInfoOpen).to.be.true;
+      expect(orderEditPage.renewalsAccordion.renuvalInterval).to.be.true;
     });
   });